pyproject.toml
[project]
name = "sportswear-esg-news-classifier"
version = "0.1.0"
description = "ESG News Classifier for sportswear brands - multi-label text classification for Environmental, Social, and Governance news articles"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "newsdataapi>=0.1.12",
    "sqlalchemy>=2.0",
    "psycopg2-binary>=2.9",
    "pgvector>=0.3",
    "python-dotenv>=1.0",
    "pydantic>=2.0",
    "newspaper4k>=0.9",
    "httpx>=0.27",
    "langdetect>=1.0.9",
    # Labeling pipeline dependencies
    "anthropic>=0.18",
    "openai>=1.0",
    "tiktoken>=0.5",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=4.0",
]

[project.scripts]
collect-news = "scripts.collect_news:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
